Extract remaining sections, eliminate ModalForm
