+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/components/dialogs/MoreDialog/Hint.tsx
@@ -1,0 +1,5 @@
+export const Hint = ({ children, ...props }: React.HTMLAttributes<HTMLParagraphElement>) => (
+  <p {...props} className="mb-5 leading-snug">
+    {children}
+  </p>
+);
