+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/components/dialogs/MoreDialog/ModalLabel.tsx
@@ -1,0 +1,5 @@
+export const ModalLabel = ({ children, ...props }: React.LabelHTMLAttributes<HTMLLabelElement>) => (
+  <label {...props} className="mb-2 block font-[700]">
+    {children}
+  </label>
+);
